feat: add genesis download functionality with CLI integration
- Add genesis download with SHA-256 verification for known chains
- New download-genesis CLI command
- Auto-download during init for public chains with warnings
- Support celestia, mocha-4, arabica-10, arabica-11 networks
